1-listint_len.c

+++ 0x12-more_singly_linked_lists/1-listint_len.c
@@ -5,17 +5,18 @@
  * listint_len - function that returns the number of elements in a linked
  * listint_t list.
  * @h: pointer used to iterate through the number of elements in a list.
+ * Return : number of elements in a linked list.
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (h == NULL)
 		return (0);
 	while (h != NULL)
 	{
 		h = h->next;
-	i++;
+		i++;
 	}
 
 	return (i);
